[RFR] mail: Move services and more to the core
This commit moves these files to core:
- command_palettes.js
- messaging_hook.js
- messaging_service.js
- primary_variables.scss
- sound_effects-service.js
- user_settings_service.js
