+++ addons/mail/__manifest__.py
@@ -177,7 +177,7 @@ For more specific needs, you may also assign custom-defined actions
         ],
         'web._assets_primary_variables': [
             'mail/static/src/scss/variables/primary_variables.scss',
-            'mail/static/src/new/primary_variables.scss',
+            'mail/static/src/new/core/primary_variables.scss',
         ],
         'web.assets_backend': [
             # depends on BS variables, can't be loaded in assets_primary or assets_secondary
