+++ addons/mail/static/src/new/chat/chat_window_icon.js
@@ -1,7 +1,7 @@
 /* @odoo-module */
 
 import { Component } from "@odoo/owl";
-import { useMessaging } from "../messaging_hook";
+import { useMessaging } from "../core/messaging_hook";
 import { Typing } from "../composer/typing";
 
 /**
