+++ addons/mail/static/src/new/core/command_palette.js
