+++ addons/mail/static/src/new/core/messaging_hook.js
