+++ addons/mail/static/src/new/core/primary_variables.scss
