+++ addons/mail/static/src/new/core/sound_effects_service.js
