+++ addons/mail/static/src/new/core/user_settings_service.js
