+++ addons/mail/static/src/new/messaging_menu/messaging_menu.js
@@ -2,7 +2,7 @@
 
 import { Component, useState } from "@odoo/owl";
 import { Dropdown } from "@web/core/dropdown/dropdown";
-import { useMessaging } from "../messaging_hook";
+import { useMessaging } from "../core/messaging_hook";
 import { PartnerImStatus } from "@mail/new/discuss/partner_im_status";
 import { NotificationItem } from "./notification_item";
 import { browser } from "@web/core/browser/browser";
