+++ addons/mail/static/src/new/rtc/call_settings.js
@@ -1,7 +1,7 @@
 /* @odoo-module */
 
 import { Component } from "@odoo/owl";
-import { useMessaging } from "../messaging_hook";
+import { useMessaging } from "../core/messaging_hook";
 
 export class CallSettings extends Component {
     static template = "mail.settings";
