+++ addons/mail/static/tests/helpers/webclient_setup.js
@@ -13,7 +13,7 @@ import { PopoverManagerContainer } from "@mail/components/popover_manager_contai
 import { ActivityMenu } from "@mail/new/activity/activity_menu";
 import { ChatWindowContainer } from "@mail/new/chat/chat_window_container";
 import { MessagingMenu } from "@mail/new/messaging_menu/messaging_menu";
-import { messagingService as newMessagingService } from "@mail/new/messaging_service";
+import { messagingService as newMessagingService } from "@mail/new/core/messaging_service";
 import { attachmentViewerService } from "@mail/new/attachment_viewer/attachment_viewer_service";
 import { messagingService } from "@mail/services/messaging_service";
 import { systrayService } from "@mail/services/systray_service";
@@ -24,8 +24,8 @@ import { registry } from "@web/core/registry";
 import { patchWithCleanup } from "@web/../tests/helpers/utils";
 import { createWebClient } from "@web/../tests/webclient/helpers";
 import { effectService } from "@web/core/effects/effect_service";
-import { soundEffects } from "@mail/new/sound_effects_service";
-import { userSettingsService } from "@mail/new/user_settings_service";
+import { soundEffects } from "@mail/new/core/sound_effects_service";
+import { userSettingsService } from "@mail/new/core/user_settings_service";
 import { rtcService } from "@mail/new/rtc/rtc_service";
 import { suggestionService } from "@mail/new/suggestion/suggestion_service";
 import { stateService } from "@mail/new/core/state_service";
